HW21 refactoring
