+++ Homework21/Homework21/Homework21.csproj
@@ -57,7 +57,10 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Program.cs" />
+    <Compile Include="ConsoleLogger.cs" />
+    <Compile Include="DataManager.cs" />
+    <Compile Include="IDataLogger.cs" />
+    <Compile Include="StudentChecker.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
